refactor(face): merge tw & wt action modules into widget-twidget, rename renderer vars, and simplify TWidgetFaceProps
