+++ src/entities/face/api/actions/index.ts
@@ -1,5 +1,4 @@
 export * from "./face";
 export * from "./nb";
 export * from "./pg";
-export * from "./tw";
-export * from "./wt";
+export * from "./widget-twidget";
